+++ src/Symfony/Component/Messenger/EventListener/StopWorkerOnIdleListener.php
@@ -1,0 +1,45 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Messenger\EventListener;
+
+use Psr\Log\LoggerInterface;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\Messenger\Event\WorkerRunningEvent;
+
+/**
+ * Stops the worker as soon as its receivers return no message.
+ *
+ * The worker handles every message that is available and exits without sleeping.
+ */
+class StopWorkerOnIdleListener implements EventSubscriberInterface
+{
+    public function __construct(
+        private ?LoggerInterface $logger = null,
+    ) {
+    }
+
+    public function onWorkerRunning(WorkerRunningEvent $event): void
+    {
+        if ($event->isWorkerIdle()) {
+            $event->getWorker()->stop();
+
+            $this->logger?->info('Worker stopped because no message was available');
+        }
+    }
+
+    public static function getSubscribedEvents(): array
+    {
+        return [
+            WorkerRunningEvent::class => 'onWorkerRunning',
+        ];
+    }
+}